revela-app/pyproject.toml
[project]
name = "revela"
version = "0.1.0"
description = "Right-click any chart or table on the web to get instant AI insights, powered by Gemma on Cloud Run"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "google-auth>=2.23.0",
    "requests>=2.31.0",
    "gunicorn>=23.0.0",
    "flask-cors>=6.0.1",
    "polars>=1.13.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
]
